requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
black==21.12b0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
chromedriver-autoinstaller==0.3.1
click==8.1.3
colorclass==2.2.2
cryptography==37.0.4
docopt==0.6.2
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
idna==3.3
mypy-extensions==0.4.3
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
pathspec==0.9.0
pip-upgrader==1.4.15
platformdirs==2.5.2
protobuf==4.21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
terminaltables==3.1.10
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
webdriver-manager==3.8.2
wsproto==1.1.0
